refactor: reorganize add-blocks components and update imports for consistency
